replace settings_bootstrap folder name with settings
